beginning web services. added ASP.NET project

+++ Kung Fu Tracker/Kung_Fu_Tracker/Models/User.cs
@@ -14,7 +14,7 @@ namespace Kung_Fu_Tracker.Models
         public string Username { get; set; }
         public string Password { get; set; }
         public string Rank { get; set; }
-        public byte Stripes { get; set; }
+        public int Stripes { get; set; }
 
         public User() { }
         public User(string name, string pass)
